api/api_bicycle_routes/requirements.txt
asgiref==3.5.2
Cerberus==1.3.4
Django==4.1.3
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
